Move worker to library
