+++ lading_generators/src/splunk_hec_gen.rs
@@ -3,4 +3,6 @@
 //! This crate is intended to back the `splunk_hec_gen` executable and is
 //! not considered useful otherwise.
 
+pub use worker::Worker;
 pub mod config;
+mod worker;
